backend/requirements.txt
fastapi
uvicorn
pillow
google-generativeai
python-multipart
